+++ src/components/Login/Login.tsx
@@ -2,6 +2,7 @@ import './Login.css';
 import imagenLogin from '../../images/imagenVioleta.png';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import { handleLogin } from './servicioLogin'; // Asegúrate de ajustar la ruta de importación según tu estructura de carpetas
 
 const Login: React.FC = () => {
   const [email, setEmail] = useState<string>('');
@@ -9,34 +10,9 @@ const Login: React.FC = () => {
   const [error, setError] = useState<string | null>(null);
   const navigate = useNavigate();
 
-  const handleLogin = async (e: React.FormEvent<HTMLFormElement>) => {
+  const onSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-
-    const data = {
-      email: email,
-      password: password,
-    };
-
-    try {
-      const response = await fetch('http://localhost:3000/teachers/by-email', { // Ajusta esta URL a tu endpoint real
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(data),
-      });
-
-      if (!response.ok) {
-        throw new Error('Invalid credentials');
-      }
-
-      const result = await response.json();
-      console.log('Login successful:', result);
-      navigate('/dashboard'); // Redirige al usuario al dashboard
-    } catch (error) {
-      setError('Invalid email or password');
-      console.error('Error:', error);
-    }
+    await handleLogin(email, password, setError, navigate);
   };
 
   return (
@@ -47,7 +23,7 @@ const Login: React.FC = () => {
       <div className="login-form">
         <h2>Sign In</h2>
         {error && <div className="error-message">{error}</div>}
-        <form onSubmit={handleLogin}>
+        <form onSubmit={onSubmit}>
           <div className="input-group">
             <label htmlFor="email">Email :</label>
             <div className="input-wrapper">
