+++ src/home/Home.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { useParams } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import { getDatos, fetchData } from "./data";
 import ListStudents from "./ListStudents";
 import SearchStudents from "./SearchStudents";
@@ -7,17 +7,21 @@ import { DataType } from "./types";
 import Menu from "./Menu";
 
 const Home: React.FC = () => {
-  const { teacherID } = useParams<{ teacherID: string }>();
+  const navigate = useNavigate();
   const [searchText, setSearchText] = useState<string>("");
   const [datos, setDatos] = useState<DataType[]>([]);
+  const teacherID = localStorage.getItem('teacherID');
 
   useEffect(() => {
-    if (teacherID) {
-      fetchData(Number(teacherID)).then(() => {
-        setDatos(getDatos());
-      });
+    if (!teacherID) {
+      navigate('/login'); // Redirige al usuario al login si no hay teacherID en localStorage
+      return;
     }
-  }, [teacherID]);
+
+    fetchData(Number(teacherID)).then(() => {
+      setDatos(getDatos());
+    });
+  }, [teacherID, navigate]);
 
   const handleSearch = (value: string) => {
     setSearchText(value);
@@ -32,7 +36,7 @@ const Home: React.FC = () => {
   return (
     <div className="flex flex-col h-screen bg-stone-300">
       <div className="flex justify-between p-4">
-        <Menu TeacherID={teacherID}/>
+        <Menu TeacherID={parseInt(teacherID || '0')} />
       </div>
       <div className="flex justify-center items-center flex-grow">
         <div className="w-2/3 min-h-[809px]">
